Move `urls` package to `pkg` directory
We try to use the `cmd/uhc` directory only for the implementations of
the commands. All other functions and types should, if possible, go in
the `pkg` directory. This patch moves the `urls` package to stick to
that convention.
