+++ pkg/urls/url_expander.go
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2018 Red Hat, Inc.
+Copyright (c) 2019 Red Hat, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
